scripts/requirements.txt
pandas==2.1.1
protobuf==4.24.4
SQLAlchemy==2.0.22
aiohttp==3.8.6
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloud-sql-python-connector==1.4.2
cryptography==41.0.4
docopt==0.6.2
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
googlemaps==4.10.0
greenlet==3.0.0
idna==3.4
multidict==6.0.4
numpy==1.26.0
pandas==2.1.1
pg8000==1.30.2
pipreqs==0.4.13
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
yarg==0.1.9
yarl==1.9.2
